Update RaceUpdate to use new state entity
We no longer need payload files as we've moved to an entities interface
that allows us to better navigate through different types of updates
from the client.
